move some functions, classes, and others from both's main command file to those base files.
