+++ src/commands/redmine.ts
@@ -107,133 +107,4 @@ ${obj.description}
       this.error(`${e.message}`)
     }
   }
-
-  readonly ValidationFlags = (flags: any) => {
-    if (isNaN(parseInt(flags.project, 10))) {
-      throw new Error('project id(-p, --project) is not a number')
-    }
-    if (isNaN(parseInt(flags.ticket, 10))) {
-      throw new Error('ticket id(-t, --ticket) is not a number')
-    }
-    if (isNaN(parseInt(flags.limit, 10))) {
-      throw new Error('query id(-l, --limit) is not a number')
-    }
-    if (isNaN(parseInt(flags.offset, 10))) {
-      throw new Error('query id(-o, --offset) is not a number')
-    }
-  }
-
-  readonly getProjectId = async (rApi: RedmineApi) => {
-    try {
-      const projects = await rApi.get(rApi.getProjectsURL(), rApi.createParams(null, null, 100))
-      const argsProjects = projects.data.projects.map((obj: any) => {
-        return {name: obj.name, value: obj.id}
-      })
-      const projectsList: object = {
-        name: 'id',
-        message: 'choose project',
-        type: 'list',
-        choices: argsProjects
-      }
-      const choosed: any = await this.inquiry(projectsList)
-      return parseInt(choosed.id, 10)
-    } catch (e) {
-      throw new Error('Fail get projects')
-    }
-  }
-
-  readonly getIssueStatusId = async (rApi: RedmineApi) => {
-    try {
-      const issueStatuses = await rApi.get(rApi.getIssueStatusesURL(), rApi.createParams(null, null, 100))
-      const argsIssueStatuses: Array<object> = issueStatuses.data.issue_statuses.map((obj: any) => {
-        return {name: obj.name, value: obj.id}
-      })
-      argsIssueStatuses.unshift({id: null, name: '指定なし'})
-      const issuesStatusesList: object = {
-        name: 'id',
-        message: 'choose issues status',
-        type: 'list',
-        choices: argsIssueStatuses
-      }
-      const choosed: any = await this.inquiry(issuesStatusesList)
-      return parseInt(choosed.id, 10)
-    } catch (e) {
-      this.warn('Fail get issue statuses')
-      return null
-    }
-  }
-
-  readonly getIssueTrackerrId = async (rApi: RedmineApi) => {
-    try {
-      const issueTrackers = await rApi.get(rApi.getIssueTrackersURL(), rApi.createParams(null, null, 100))
-      if (issueTrackers.data.trackers.length === 0) {
-        this.log('this project has no tracker')
-        return null
-      }
-      const argsIssueTrackers: Array<object> = issueTrackers.data.trackers.map((obj: any) => {
-        return {name: obj.name, value: obj.id}
-      })
-      argsIssueTrackers.unshift({id: null, name: '指定なし'})
-      const issuesTrackersList: object = {
-        name: 'id',
-        message: 'choose issues trakcer',
-        type: 'list',
-        choices: argsIssueTrackers
-      }
-      const choosed: any = await this.inquiry(issuesTrackersList)
-      return parseInt(choosed.id, 10)
-    } catch (e) {
-      this.warn('Fail get issue trakcer')
-      return null
-    }
-  }
-
-  readonly getIssueCategoryId = async (rApi: RedmineApi, projectId: number) => {
-    try {
-      const issueCategories = await rApi.get(rApi.getIssueCategoriesURL(String(projectId)), rApi.createParams(null, null, 100))
-      if (issueCategories.data.issue_categories.length === 0) {
-        this.log('this project has no categories')
-        return null
-      }
-      const argsIssueCategories: Array<object> = issueCategories.data.issue_categories.map((obj: any) => {
-        return {name: obj.name, value: obj.id}
-      })
-      argsIssueCategories.unshift({id: null, name: '指定なし'})
-      const issuesCategoriesList: object = {
-        name: 'id',
-        message: 'choose issues status',
-        type: 'list',
-        choices: argsIssueCategories
-      }
-      const choosed: any = await this.inquiry(issuesCategoriesList)
-      return parseInt(choosed.id, 10)
-    } catch (e) {
-      this.warn('Fail get issue category')
-      return null
-    }
-  }
-
-  readonly getQueryId = async (rApi: RedmineApi, pid: number) => {
-    const queries = await rApi.get(rApi.getQueriesURL(), rApi.createParams(null, null, 100))
-    const argsQueries = queries.data.queries.map((obj: any) => {
-      if (pid === obj.project_id) {
-        return {name: obj.name, value: obj.id}
-      } else {
-        return false
-      }
-    }).filter(Boolean)
-    if (argsQueries.length === 0) {
-      this.warn(`not exists queries in project`)
-      this.log(`show all tickets in project`)
-      return null
-    }
-    const queriesList: object = {
-      name: 'id',
-      message: 'choose query',
-      type: 'list',
-      choices: argsQueries
-    }
-    const choosed: any = await this.inquiry(queriesList)
-    return parseInt(choosed.id, 10)
-  }
 }
